removes front ends from scala-reflect.jar
It was an interesting idea to give macro developers control over front ends,
but it hasn't given any visible results.

To the contrast, front ends have proven useful for toolboxes to easily control
what errors get printed where.

Therefore I'm moving front ends to scala-compiler.jar to clean up the API.
Yay for scaladoc-driven development!
