+++ test/files/run/toolbox_console_reporter.scala
@@ -1,16 +1,29 @@
 import scala.reflect.runtime.universe._
+import scala.reflect.runtime.{universe => ru}
+import scala.reflect.runtime.{currentMirror => cm}
+import scala.tools.reflect.{ToolBox, mkConsoleFrontEnd}
 
 object Test extends App {
-  // todo. cannot test this unfortunately, because ConsoleFrontEnd grabs Console.out too early
-  // todo. and isn't affected by Console.setOut employed by partest to intercept output
+  val oldErr = Console.err;
+  val baos = new java.io.ByteArrayOutputStream();
+  Console.setErr(new java.io.PrintStream(baos));
+  try {
+    val toolbox = cm.mkToolBox(frontEnd = mkConsoleFrontEnd(), options = "-deprecation")
+    toolbox.eval(reify{
+      object Utils {
+        @deprecated("test", "2.10.0")
+        def foo { println("hello") }
+      }
 
-  //val toolbox = mkToolBox(frontEnd = mkConsoleFrontEnd(), options = "-deprecation")
-  //toolbox.eval(reify{
-  //  object Utils {
-  //    @deprecated("test", "2.10.0")
-  //    def foo { println("hello") }
-  //  }
-  //
-  //  Utils.foo
-  //})
+      Utils.foo
+    }.tree)
+    println("============compiler console=============")
+    println(baos.toString);
+    println("=========================================")
+    println("============compiler messages============")
+    toolbox.frontEnd.infos.foreach(println(_))
+    println("=========================================")
+  } finally {
+    Console.setErr(oldErr);
+  }
 }
