+++ test/files/run/toolbox_silent_reporter.scala
@@ -1,10 +1,10 @@
 import scala.reflect.runtime.universe._
 import scala.reflect.runtime.{universe => ru}
 import scala.reflect.runtime.{currentMirror => cm}
-import scala.tools.reflect.ToolBox
+import scala.tools.reflect.{ToolBox, mkSilentFrontEnd}
 
 object Test extends App {
-  val toolbox = cm.mkToolBox(options = "-deprecation")
+  val toolbox = cm.mkToolBox(options = "-deprecation", frontEnd = mkSilentFrontEnd())
   toolbox.eval(reify{
     object Utils {
       @deprecated("test", "2.10.0")
